refactor(tests): extract shared rdflib shadow test helper
Move the duplicated "_run_in_shadow" helper and the shadow directory constant into a dedicated test utility module so both test_rdflib_shadow.py and test_term_model.py share a single implementation.
